Support for different deck draw sizes

+++ LoRSideTracker/Core/CardLibrary.cs
@@ -16,7 +16,8 @@ namespace LoRSideTracker
     /// </summary>
     public static class CardLibrary
     {
-        private static Dictionary<string, Card> m_cards;
+        private static Dictionary<string, Card> Cards;
+        private static Dictionary<string, Region> Regions;
 
         /// <summary>
         /// Find the list of missing sets that can be downloaded
@@ -29,7 +30,7 @@ namespace LoRSideTracker
                 Directory.CreateDirectory(Constants.GetLocalSetsPath());
             }
             var missingSets = new List<ValueTuple<int, long>>();
-            for (int i = 1; i < 99; i++)
+            for (int i = 0; i < 99; i++)
             {
                 long remoteZipSize = Utilities.CheckURLExists(Constants.GetSetURL(i));
                 if (remoteZipSize <= 0)
@@ -100,27 +101,30 @@ namespace LoRSideTracker
             File.WriteAllText(Constants.GetSetVersionInfoPath(setNumber), json);
 
             // Resize all images to manageable size
-            string imagesDir = String.Format("{0}\\{1}\\img\\cards", setPath, Constants.Language);
-            DirectoryInfo dirInfo = new DirectoryInfo(imagesDir);
-            FileInfo[] files = dirInfo.GetFiles();
-            const int targetWidth = 256;
-            foreach (FileInfo fi in files)
+            if (setNumber > 0)
             {
-                Image fullSizeImage = Image.FromFile(fi.FullName);
-                double ratio = (double)targetWidth / (double)fullSizeImage.Width;
+                string imagesDir = String.Format("{0}\\img\\cards", setPath);
+                DirectoryInfo dirInfo = new DirectoryInfo(imagesDir);
+                FileInfo[] files = dirInfo.GetFiles();
+                const int targetWidth = 256;
+                foreach (FileInfo fi in files)
+                {
+                    Image fullSizeImage = Image.FromFile(fi.FullName);
+                    double ratio = (double)targetWidth / (double)fullSizeImage.Width;
 
-                if (ratio < 1.0)
-                {
-                    int targetHeight = (int)(0.5 + fullSizeImage.Height * ratio);
-                    Image smallImage = Utilities.ResizeImage(fullSizeImage, targetWidth, targetHeight);
-                    fullSizeImage.Dispose();
-                    File.Delete(fi.FullName);
-                    smallImage.Save(fi.FullName, System.Drawing.Imaging.ImageFormat.Png);
-                    smallImage.Dispose();
-                }
-                else
-                {
-                    fullSizeImage.Dispose();
+                    if (ratio < 1.0)
+                    {
+                        int targetHeight = (int)(0.5 + fullSizeImage.Height * ratio);
+                        Image smallImage = Utilities.ResizeImage(fullSizeImage, targetWidth, targetHeight);
+                        fullSizeImage.Dispose();
+                        File.Delete(fi.FullName);
+                        smallImage.Save(fi.FullName, System.Drawing.Imaging.ImageFormat.Png);
+                        smallImage.Dispose();
+                    }
+                    else
+                    {
+                        fullSizeImage.Dispose();
+                    }
                 }
             }
             File.Delete(setZip);
@@ -132,20 +136,19 @@ namespace LoRSideTracker
         /// <param name="pd">ProgressDisplay interface to use for progress updates</param>
         public static void LoadAllCards(ProgressDisplayControl pd)
         {
-            m_cards = new Dictionary<string, Card>();
+            Cards = new Dictionary<string, Card>();
             for (int i = 1; ; i++)
             {
-                string setName = Constants.GetSetName(i);
-                string setDir = String.Format("{0}\\{1}\\{2}\\", Constants.GetLocalSetsPath(), setName, Constants.Language);
-                string setJsonFile = String.Format("{0}\\data\\{1}.json", setDir, setName);
+                string setDir = Constants.GetSetPath(i);
+                string setJsonFile = String.Format("{0}\\data\\{1}.json", setDir, Constants.GetSetName(i));
                 if (!File.Exists(setJsonFile))
                 {
                     // Done
-                    return;
+                    break;
                 }
 
-                string json = Utilities.ReadLocalFile(setJsonFile);
-                Dictionary<string, JsonElement>[] set = JsonSerializer.Deserialize<Dictionary<string, JsonElement>[]>(json);
+                string coreJson = Utilities.ReadLocalFile(setJsonFile);
+                Dictionary<string, JsonElement>[] set = JsonSerializer.Deserialize<Dictionary<string, JsonElement>[]>(coreJson);
 
                 for (int j = 0; j < set.Length; j++)
                 {
@@ -155,9 +158,23 @@ namespace LoRSideTracker
                         pd.Update(String.Format("Processing set {0}, card {1} of {2}.", i, j + 1, set.Length), percentage);
                     }
                     Dictionary<string, JsonElement> dict = set[j];
-                    m_cards.Add(dict["cardCode"].ToString(), new Card(setDir, dict));
+                    Cards.Add(dict["cardCode"].ToString(), new Card(setDir, dict));
                 }
             }
+
+            // Load all regions
+            string coreDir = Constants.GetSetPath(0);
+            string coreJsonFile = String.Format("{0}\\data\\globals-en_us.json", coreDir);
+            string json = Utilities.ReadLocalFile(coreJsonFile);
+            Dictionary<string, JsonElement> core = JsonSerializer.Deserialize<Dictionary<string, JsonElement>>(json);
+            Dictionary<string, JsonElement>[] regions = core["regions"].ToObject<Dictionary<string, JsonElement>[]>();
+            Regions = new Dictionary<string, Region>();
+            foreach (var region in regions)
+            {
+                Region r = new Region(coreDir, region);
+                Regions[r.NameRef] = r;
+            }
+
         }
 
         /// <summary>
@@ -169,11 +186,28 @@ namespace LoRSideTracker
         {
             try
             {
-                return m_cards[cardCode];
+                return Cards[cardCode];
             }
             catch
             {
                 return Card.UnknownCard;
+            }
+        }
+
+        /// <summary>
+        /// Get the region corresponding to the region name
+        /// </summary>
+        /// <param name="regionNameRef">Region name</param>
+        /// <returns>Corresponding Region object</returns>
+        public static Region GetRegion(string regionNameRef)
+        {
+            try
+            {
+                return Regions[regionNameRef];
+            }
+            catch
+            {
+                return Region.UnknownRegion;
             }
         }
     }
